extract mix-specific adapter and tests
